+++ apps/ui-dev/constants/design-pages.js
@@ -2,6 +2,8 @@ export default [
   { link: "/pages/design-system/", title: "Getting Started" },
   "Design Tokens",
   { link: "/pages/design-system/color/", title: "Color" },
+  { link: "/pages/design-system/states/", title: "States" },
+  { link: "/pages/design-system/icongraphy/", title: "Icongraphy" },
   { link: "/pages/design-system/typography/", title: "Typography" },
   { link: "/pages/design-system/spacing/", title: "Spacing" },
   { link: "/pages/design-system/layout/", title: "Layout" },