+++ apps/ui-dev/pages/ui-component/data-editor.jsx
@@ -1,4 +1,4 @@
-import AddFilled from "@elf-framework/icon/esm/AddFilled";
+import AddFilled from "@elf-framework/icon/AddFilled";
 import { start, SUBSCRIBE, UIElement, useSubscribe } from "@elf-framework/sapa";
 import { DataEditor, View } from "@elf-framework/ui";
 import "@elf-framework/ui/style.css";